Remove unnecessary stuff
This change created a single spec failure only caused the first time
specs are run when the test gems installed in `tmp/gems` have not yet
being installed.

The problem was that the code branch that installs the gems uses the
`FileUtils` constant. Accessing that constant makes [this autoload]
trigger the custom rubygems `require` behavior which will end up setting
`gem_home` to its current value (when calling `to_spec` on the selected
default gem). At that point, even if the `Gem.clear_paths` call has already
happened, but the old value for `Gem.home` is restored and memoized.

That means when the tests updates `ENV["HOME"]`, it has no effect in the path
to the configuration file rubygems uses, and the spec checking that invalid
YAML in the configuration file gives a friendly error will fail, because the
file is not even loaded.

This could be fixed by removing the `FileUtils` autoload, or stop
memoizing `Gem.home`, but the simplest fix seemed to require `fileutils`
early, so that `Gem.home` is not memoized again after calling
`Gem.clear_paths`.

[this autoload]: https://github.com/rubygems/rubygems/blob/511a0f5105ca95b8e389bf477b6c7cf3e90be7d1/lib/rubygems/source.rb#L3